* test/pyhttpd/env.py (HttpdTestEnv.apache_hard_restart, read_pid_file):   Moved here from the mod_systemd tests, for use by other suites.
No functional change.
